backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Todo Application Backend - Phase II"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlmodel>=0.0.16",
    "better-auth>=0.0.0-beta.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.9",
    "PyJWT>=2.8.0",
    "psycopg2-binary>=2.9.7",
    "openai>=1.0.0"
]

[project.scripts]
dev = "main:dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.25.0"
]